Fix __init__.py so deploy can be imported
Moves all code from __init__.py to cli.py

Closes #35
