+++ setup.cfg
@@ -37,7 +37,7 @@ include_package_data=True
 
 [options.entry_points]
 console_scripts =
-    phantom = phtoolbox:main
+    phantom = phtoolbox.cli:main
 
 [options.packages.find]
 where = src
