+++ src/phtoolbox/__main__.py
@@ -1,5 +1,5 @@
 # This file is needed to allow python -m to work.
-from phtoolbox import main
+from phtoolbox.cli import main
 
 if __name__ == "__main__":
     main()
